Moving MapConfig to separate file
